+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `SplitFileMerger` to `pyctr.fileio` to merge multiple file-like objects into one (currently no support for writing)
 * Support closing all subfiles in `SplitFileMerger`
 * Rewrite ExeFS handling in `NCCHReader` to use `SplitFileMerger` to merge multiple `SubsectionIO` files to handle the parts that use different encryption, and update `FullDecrypted` to use it when reading ExeFS
+* Add `from_bytes` classmethod to `NCCHFlags`
 
 ## v0.4.7 - April 20, 2021
 * Use absolute paths in `CDNReader`